+++ app/src/main/java/com/mprusina/gitrepo/repos/presenter/RepoListPresenter.kt
@@ -9,9 +9,7 @@ import com.mprusina.gitrepo.common.api.GitHubService
 import com.mprusina.gitrepo.common.model.Repo
 import com.mprusina.gitrepo.repos.ReposContract
 import com.mprusina.gitrepo.repos.data.RepoPagingSource
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.runBlocking
@@ -22,10 +20,6 @@ private const val REPO_PAGE_SIZE = 50
 class RepoListPresenter @Inject constructor(private val dbRepository: DatabaseRepository, private val service: GitHubService) : ReposContract.Presenter {
 
     private lateinit var repoPagingData : Flow<PagingData<Repo>>
-
-    init {
-        loadData()
-    }
 
     override fun loadData(): Flow<PagingData<Repo>> {
         repoPagingData = Pager(
@@ -42,27 +36,20 @@ class RepoListPresenter @Inject constructor(private val dbRepository: DatabaseRe
         val repoSearch = repoPagingData
         return repoSearch.map {
             it.filter { repo ->
-                repo.name?.contains(query) == true
-                        || repo.owner?.contains(query) == true
-                        || repo.description?.contains(query) == true
-            }
-        }.flowOn(Dispatchers.Default)
-    }
-
-    override fun saveToFavorites(repo: Repo) {
-        repo.favorite = true
-        runBlocking {
-            launch(Dispatchers.IO) {
-                dbRepository.saveRepoToFavorites(repo)
+                repo.name?.contains(query) == true || repo.owner?.contains(query) == true || repo.description?.contains(query) == true
             }
         }
     }
 
-    override fun removeFromFavorites(repo: Repo) {
-        repo.favorite = false
+    override fun handleRepoFavoriteAction(repo: Repo) {
+        repo.favorite = repo.favorite != true
         runBlocking {
-            launch(Dispatchers.IO) {
-                dbRepository.deleteRepoFromFavorites(repo)
+            launch {
+                if (repo.favorite == true) {
+                    dbRepository.saveRepoToFavorites(repo)
+                } else {
+                    dbRepository.deleteRepoFromFavorites(repo)
+                }
             }
         }
     }
